+++ packages/react/CHANGELOG.md
@@ -1,5 +1,64 @@
 # @uploadthing/react
 
+## 5.2.1
+
+### Patch Changes
+
+- [#251](https://github.com/pingdotgg/uploadthing/pull/251)
+  [`872c8a0`](https://github.com/pingdotgg/uploadthing/commit/872c8a08e01cd2c0f59a837b410bf4b0fc29ce9c)
+  Thanks [@markflorkowski](https://github.com/markflorkowski)! - fix: coherent
+  file info in all methods
+
+  all methods now receives a similarly shaped object as the serverside
+  `onUploadComplete` callback:
+
+  ```ts
+  export type UploadFileResponse = {
+    name: string;
+    size: number;
+    key: string;
+    url: string;
+  };
+  ```
+
+  Updated methods are:
+
+  - `onClientUploadComplete` in hooks as well as components (The old `fileName`,
+    `fileSize`, `fileUrl` and `fileKey` are retained but marked as deprecated
+    for backwards compatibility, and will be removed in the next major.)
+  - `utapi.uploadFiles` as well as `utapi.uploadFilesFromUrl`
+
+- Updated dependencies
+  [[`fe46b81`](https://github.com/pingdotgg/uploadthing/commit/fe46b814aa75646eac0694fdcb3889a3f7f5122b)]:
+  - @uploadthing/shared@5.2.0
+
+## 5.2.0
+
+### Minor Changes
+
+- [#195](https://github.com/pingdotgg/uploadthing/pull/195)
+  [`a6c969e`](https://github.com/pingdotgg/uploadthing/commit/a6c969e67c85df490907b121d8e7df41779172b3)
+  Thanks [@juliusmarminge](https://github.com/juliusmarminge)! - feat: improve
+  errors and add `errorFormatter` option on the backend
+
+### Patch Changes
+
+- [#205](https://github.com/pingdotgg/uploadthing/pull/205)
+  [`8658002`](https://github.com/pingdotgg/uploadthing/commit/8658002ca01e6502f06c2c56f90b353cf5db71df)
+  Thanks [@juliusmarminge](https://github.com/juliusmarminge)! - chore(deps):
+  update dependency '@uploadthing/shared'
+
+- [#207](https://github.com/pingdotgg/uploadthing/pull/207)
+  [`f3640fb`](https://github.com/pingdotgg/uploadthing/commit/f3640fb0872a12fe0ad95e0ac604136113fa6ca4)
+  Thanks [@markflorkowski](https://github.com/markflorkowski)! - fix(shared):
+  Swap order of cases in `getUploadthingUrl()`
+
+- Updated dependencies
+  [[`c508868`](https://github.com/pingdotgg/uploadthing/commit/c508868690d3de48094c10a7facb880662d03b6a),
+  [`f3640fb`](https://github.com/pingdotgg/uploadthing/commit/f3640fb0872a12fe0ad95e0ac604136113fa6ca4),
+  [`a6c969e`](https://github.com/pingdotgg/uploadthing/commit/a6c969e67c85df490907b121d8e7df41779172b3)]:
+  - @uploadthing/shared@5.1.0
+
 ## 5.1.0
 
 ### Minor Changes